Fix timestamp normalization math, fixing rare timestamp bug.
Occasionally the packet forwarder will forward a packet with
an unparsable `time` timestamp attribute because the fractional
part of the seconds field will have a negative sign in it.

An example that I witnessed was the timestamp:

`2020-10-11T22:59:05.-00072Z`

Bugs like this are almost always signs of buggy attempts to
perform arithmetic on `timeval` structures in C without carefully
covering all of the overflow _and_ underflow conditions. Sure
enough, that was the case here.

With this change, the code now properly handles a `tv_nsec`
underflow and should no longer cause other parts of the code to
generate unparsable timestamps like the given example.